fix: (poll block) Windows visible event triggers rerender app every time
